Moved deck migration runnable to it's own package
